+++ src/ext-test/java/org/opentripplanner/ext/SandboxExtensionResourceTest.java
@@ -1,16 +1,17 @@
 package org.opentripplanner.ext;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.junit.Test;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.LineNumberReader;
-import org.junit.jupiter.api.Test;
+
+import static org.junit.Assert.assertTrue;
 
 /**
  * This test has NOTHING to do with the OJP extension. The test just make
- * sure that the Maven project is correctly set up and that the 'ext-test/resources'
+ * sure that the Maven project is correctly set up amd that the 'ext-test/resources'
  * is available on classpath.
  */
 public class SandboxExtensionResourceTest {
@@ -25,6 +26,6 @@ public class SandboxExtensionResourceTest {
 
         String text = in.readLine();
 
-        assertTrue(text.length() > 2, text);
+        assertTrue(text,text.length() > 2);
     }
 }