+++ src/ext/java/org/opentripplanner/ext/flex/trip/UnscheduledTrip.java
@@ -1,6 +1,5 @@
 package org.opentripplanner.ext.flex.trip;
 
-import java.util.Set;
 import org.opentripplanner.ext.flex.FlexServiceDate;
 import org.opentripplanner.ext.flex.flexpathcalculator.FlexPathCalculator;
 import org.opentripplanner.ext.flex.template.FlexAccessTemplate;
@@ -32,8 +31,7 @@ import static org.opentripplanner.model.PickDrop.NONE;
  * trip is possible.
  */
 public class UnscheduledTrip extends FlexTrip {
-  // unscheduled trips can contain one or two stop_times
-  private static final Set<Integer> N_STOPS = Set.of(1,2);
+  private static final int N_STOPS = 2;
 
   private final UnscheduledStopTime[] stopTimes;
 
@@ -44,7 +42,7 @@ public class UnscheduledTrip extends FlexTrip {
     Predicate<StopTime> noExplicitTimes = Predicate.not(st -> st.isArrivalTimeSet() || st.isDepartureTimeSet());
     Predicate<StopTime> notContinuousStop = stopTime ->
         stopTime.getFlexContinuousDropOff() == NONE.getGtfsCode() && stopTime.getFlexContinuousPickup() == NONE.getGtfsCode();
-    return N_STOPS.contains(stopTimes.size())
+    return stopTimes.size() == N_STOPS
         && stopTimes.stream().allMatch(noExplicitTimes)
         && stopTimes.stream().allMatch(notContinuousStop);
   }
@@ -56,12 +54,11 @@ public class UnscheduledTrip extends FlexTrip {
       throw new IllegalArgumentException("Incompatible stopTimes for unscheduled trip");
     }
 
-    var size = stopTimes.size();
-    this.stopTimes = new UnscheduledStopTime[size];
-    this.dropOffBookingInfos = new BookingInfo[size];
-    this.pickupBookingInfos = new BookingInfo[size];
+    this.stopTimes = new UnscheduledStopTime[N_STOPS];
+    this.dropOffBookingInfos = new BookingInfo[N_STOPS];
+    this.pickupBookingInfos = new BookingInfo[N_STOPS];
 
-    for (int i = 0; i < size; i++) {
+    for (int i = 0; i < N_STOPS; i++) {
       this.stopTimes[i] = new UnscheduledStopTime(stopTimes.get(i));
       this.dropOffBookingInfos[i] = stopTimes.get(0).getDropOffBookingInfo();
       this.pickupBookingInfos[i] = stopTimes.get(0).getPickupBookingInfo();
@@ -73,14 +70,14 @@ public class UnscheduledTrip extends FlexTrip {
       NearbyStop access, FlexServiceDate date, FlexPathCalculator calculator, FlexParameters params
   ) {
     int fromIndex = getFromIndex(access);
-    int toIndex = getToIndex(access);
 
-    if (stopTimes[fromIndex].dropOffType == NONE.getGtfsCode()) { return Stream.empty(); }
+    if (fromIndex != 0) { return Stream.empty(); }
+    if (stopTimes[1].dropOffType == NONE.getGtfsCode()) { return Stream.empty(); }
 
     ArrayList<FlexAccessTemplate> res = new ArrayList<>();
 
-    for (StopLocation stop : expandStops(stopTimes[fromIndex].stop)) {
-      res.add(new FlexAccessTemplate(access, this, fromIndex, toIndex, stop, date, calculator, params));
+    for (StopLocation stop : expandStops(stopTimes[1].stop)) {
+      res.add(new FlexAccessTemplate(access, this, fromIndex, 1, stop, date, calculator, params));
     }
 
     return res.stream();
@@ -90,15 +87,15 @@ public class UnscheduledTrip extends FlexTrip {
   public Stream<FlexEgressTemplate> getFlexEgressTemplates(
       NearbyStop egress, FlexServiceDate date, FlexPathCalculator calculator, FlexParameters params
   ) {
-    int fromIndex = getFromIndex(egress);
     int toIndex = getToIndex(egress);
 
-    if (stopTimes[toIndex].pickupType == NONE.getGtfsCode()) { return Stream.empty(); }
+    if (toIndex != 1) { return Stream.empty(); }
+    if (stopTimes[0].pickupType == NONE.getGtfsCode()) { return Stream.empty(); }
 
     ArrayList<FlexEgressTemplate> res = new ArrayList<>();
 
-    for (StopLocation stop : expandStops(stopTimes[toIndex].stop)) {
-      res.add(new FlexEgressTemplate(egress, this, fromIndex, toIndex, stop, date, calculator, params));
+    for (StopLocation stop : expandStops(stopTimes[0].stop)) {
+      res.add(new FlexEgressTemplate(egress, this, 0, toIndex, stop, date, calculator, params));
     }
 
     return res.stream();